Rename files and update swagger
