+++ services/blockchain-connector/package.json
@@ -34,7 +34,7 @@
     "@liskhq/lisk-codec": "0.3.0-alpha.4",
     "@liskhq/lisk-cryptography": "4.0.0-alpha.4",
     "@liskhq/lisk-transactions": "6.0.0-alpha.4",
-    "@liskhq/lisk-validator": "^0.7.0-alpha.4",
+    "@liskhq/lisk-validator": "0.7.0-alpha.4",
     "big-json": "^3.1.0",
     "bluebird": "^3.7.2",
     "lisk-service-framework": "https://github.com/LiskHQ/lisk-service/raw/56e510422f755316715955754cae88a972514060/framework/dist/lisk-service-framework-1.4.7.tgz",
