+++ services/blockchain-indexer/shared/dataService/business/index.js
@@ -61,7 +61,7 @@ const {
 
 const {
 	dryRunTransactions,
-} = require('./dryRunTransactions');
+} = require('./transactionsDryRun');
 
 const {
 	getTokens,
